pyproject.toml
[tool.poetry]
name = "resume-ranking-system"
version = "1.0.0"
description = "ML Evaluation & Auditing Framework for Resume Ranking"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
sentence-transformers = "^2.2.2"
scikit-learn = "^1.3.0"
pandas = "^2.0.0"
numpy = "^1.24.0"
PyPDF2 = "^3.0.0"
pdfplumber = "^0.10.0"
rank-bm25 = "^0.2.2"
spacy = "^3.6.0"
nltk = "^3.8.0"
fastapi = {version = "^0.103.0", optional = true}
uvicorn = {version = "^0.23.0", optional = true}
streamlit = {version = "^1.26.0", optional = true}
plotly = "^5.16.0"
matplotlib = "^3.7.0"
seaborn = "^0.12.0"
textstat = "^0.7.3"
pyyaml = "^6.0"
pydantic = "^2.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
black = "^23.7.0"
flake8 = "^6.1.0"
mypy = "^1.5.0"
isort = "^5.12.0"

[tool.poetry.extras]
api = ["fastapi", "uvicorn"]
dashboard = ["streamlit"]
all = ["fastapi", "uvicorn", "streamlit"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=html --cov-report=term-missing"
